+++ amethyst_assets/src/prefab/mod.rs
@@ -352,15 +352,29 @@ where
     ) -> Result<Handle<A>, PrefabError> {
         let handle = match *self {
             AssetPrefab::Handle(ref handle) => handle.clone(),
-            AssetPrefab::File(ref name, ref format, ref options) => system_data.0.load(
+            AssetPrefab::File(..) => unreachable!(),
+        };
+        system_data.1.insert(entity, handle.clone()).map(|_| handle)
+    }
+
+    fn load_sub_assets(&mut self, progress: &mut ProgressCounter, system_data: &mut Self::SystemData) -> Result<bool, PrefabError> {
+        let handle = if let AssetPrefab::File(ref name, ref format, ref options) = *self {
+            Some(system_data.0.load(
                 name.as_ref(),
                 format.clone(),
                 options.clone(),
-                (),
+                progress,
                 &system_data.2,
-            ),
+            ))
+        } else {
+            None
         };
-        system_data.1.insert(entity, handle.clone()).map(|_| handle)
+        if let Some(handle) = handle {
+            *self = AssetPrefab::Handle(handle);
+            Ok(true)
+        } else {
+            Ok(false)
+        }
     }
 }