+++ amethyst_input/src/lib.rs
@@ -1,3 +1,6 @@
+//! A collection of abstractions for various input devices to be used with Amethyst.
+#![warn(missing_docs)]
+
 extern crate amethyst_config;
 extern crate amethyst_core;
 #[macro_use]