+++ docs/CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ### Added
 * Added base networking implementation and the `amethyst_network` crate. ([#969])
 * Support for debug lines using `DebugLines` pass, and `DebugLines` component or resource. ([#917], [#957])
-* Added JsonFormat ([#950]). 
+* Added JsonFormat ([#950]).
 * `SpriteRender` pass to draw sprites without using `Material` and `Mesh`. ([#829], [#830])
 * Sprite animation uses the `SpriteRenderChannel`. ([#829], [#830])
 * State::handle_event can now handle multiple types of events. ([#887])
@@ -31,9 +31,11 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Added AntiStorage documentation to the book. ([#997])
 * You can now stop the rotation of the FreeRotationSystem by setting HideCursor.hide value to false. ([#997])
 * Support for logging to file, toggle for logging to stdout. ([#976], [#994])
+* Added a `Hidden` Component, that hides a single entity, and a HideHierarchySystem that toggles `Hidden` on all children when used. ([#1001])
+* Documentation for drawing sprites. ([#971])
 * Added `shadow_update()` and `shadow_fixed_update()` to the `State` trait. ([#1006])
+* Added configurable width for debug lines ([#1016])
 * Added systems to pick 3d-objects from a scene with the mouse by casting camera rays. ([#1037])
-
 
 ### Changed
 * Sprites contain their dimensions and offsets to render them with the right size and desired position. ([#829], [#830])
@@ -71,6 +73,12 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Joint entities can only be part of a single skin: Materials are not swapped anymore. ([#933])
 * Fixed regression in sprite positioning after batching. ([#929])
 * Now loading default fonts from the system for UiButton ([#964])
+* Fixed single frame animation ([#1015])
+* Improved compatibility with older drivers ([#1012])
+* Forgotten `channel` field on `examples/ui` prefab ([#1024])
+* `AssetPrefab` loaded files at an incorrect time ([#1020])
+* Removed unreachable code in `TexturePrefab` ([#1020])
+* Fix OpenGL not rendering on window creation due to `glutin` bug ([#972])
 
 [#829]: https://github.com/amethyst/amethyst/issues/829
 [#830]: https://github.com/amethyst/amethyst/pull/830
@@ -99,13 +107,21 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#964]: https://github.com/amethyst/amethyst/pull/964
 [#965]: https://github.com/amethyst/amethyst/pull/965
 [#969]: https://github.com/amethyst/amethyst/pull/969
+[#971]: https://github.com/amethyst/amethyst/pull/971
+[#972]: https://github.com/amethyst/amethyst/issue/972
 [#974]: https://github.com/amethyst/amethyst/pull/974
 [#976]: https://github.com/amethyst/amethyst/pull/976
 [#981]: https://github.com/amethyst/amethyst/pull/981
 [#994]: https://github.com/amethyst/amethyst/pull/994
 [#997]: https://github.com/amethyst/amethyst/pull/997
+[#1001]: https://github.com/amethyst/amethyst/pull/1001
 [#1006]: https://github.com/amethyst/amethyst/pull/1006
 [#1008]: https://github.com/amethyst/amethyst/pull/1008
+[#1012]: https://github.com/amethyst/amethyst/pull/1012
+[#1015]: https://github.com/amethyst/amethyst/pull/1015
+[#1016]: https://github.com/amethyst/amethyst/pull/1016
+[#1024]: https://github.com/amethyst/amethyst/pull/1024
+[#1020]: https://github.com/amethyst/amethyst/pull/1020
 [winit_017]: https://github.com/tomaka/winit/blob/master/CHANGELOG.md#version-0172-2018-08-19
 [glutin_018]: https://github.com/tomaka/glutin/blob/master/CHANGELOG.md#version-0180-2018-08-03